Using BigDecimal to avoid rounding errors

+++ src/main/java/org/galatea/pocpnl/domain/Valuation.java
@@ -1,13 +1,12 @@
 package org.galatea.pocpnl.domain;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
-
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -21,14 +20,14 @@ import lombok.NoArgsConstructor;
 @Entity
 public class Valuation {
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long valuationId;
-	
-	private String book;
-	private String instrument;
-	private LocalDate date;
-	private double valuation;
-    private String inputJsonMap;
-	
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private long valuationId;
+
+  private String book;
+  private String instrument;
+  private LocalDate date;
+  private BigDecimal valuation;
+  private String inputJsonMap;
+
 }
